+++ packages/dashboard/src/components/DetailPanel/Body/index.ts
@@ -1,0 +1,2 @@
+export * from './Loading';
+export * from './MessageBanner';
